+++ doc/TDesign.Docs.Shared/Pages/Components/SelectPage.razor
@@ -5,9 +5,9 @@
     <Description></Description>
     <RunContent>
         <TSelect @bind-Value="Value" TValue="int">
-            <TSelectOption Value="1">男</TSelectOption>
-            <TSelectOption Value="2">女</TSelectOption>
-        </TSelect>
+            <TSelectOption Value="1" Label="男"/>
+            <TSelectOption Value="2" Label="女"/>
+        </TSelect> 
     </RunContent>
     <CodeContent></CodeContent>
 </Example>
