+++ src/TDesign/Components/Forms/TSelect.razor
@@ -1,0 +1,35 @@
+﻿@namespace TDesign
+@inherits BlazorInputComponentBase<TValue>
+@typeparam TValue
+
+<TPopup @ref="RefPopup" Trigger="PopupTrigger.Click" AdditionalStyle="position:absolute">
+    <ChildContent>
+        <div class="t-select__wrap">
+            <div class="t-select t-select-input">
+                <TInputText @bind-Value="SelectedLabel" Readonly SuffixIcon="@(HtmlHelper.Class.Append("t-fake-arrow").Append("t-fake-arrow--active", RefPopup.Visible).Append("t-select__right-icon").ToString())" />
+
+            </div>
+        </div>
+    </ChildContent>
+    <PopupContent>
+        <div class="t-select__dropdown-inner t-select__dropdown-inner--size-m">
+            <ul class="t-select__list">
+
+                <CascadingValue Value="this">
+
+                    @if (Options is null)
+                    {
+                        @ChildContent
+                    }
+                    else
+                    {
+                        foreach (var item in Options)
+                        {
+                            <TSelectOption Value="item.Value" Label="item.Label" />
+                        }
+                    }
+                </CascadingValue>
+            </ul>
+        </div>
+    </PopupContent>
+</TPopup>
